refactor: standardize argument formatting and adjust distance calculation types
- Updated constructor argument formatting in `NodeEvaluationContextImpl` for consistency.
- Removed squared distance calculation from `PERFORMANCE` heuristic logic.
- Changed `PathPosition.manhattanDistance` return type from `int` to `double`.

+++ api/src/main/java/de/bsommerfeld/pathetic/api/wrapper/PathPosition.java
@@ -60,7 +60,7 @@ public class PathPosition implements Cloneable {
    * @param otherPosition The other position to calculate the distance to.
    * @return The Manhattan distance between the two positions.
    */
-  public int manhattanDistance(PathPosition otherPosition) {
+  public double manhattanDistance(PathPosition otherPosition) {
     return Math.abs(this.getFlooredX() - otherPosition.getFlooredX())
         + Math.abs(this.getFlooredY() - otherPosition.getFlooredY())
         + Math.abs(this.getFlooredZ() - otherPosition.getFlooredZ());
